Improved sliding while grappling
Previously, sliding was accidentally caused by the hook, which - when the player reached it - changed the state to grappling while already grappling, which set lastTargetVelocity to zero, which caused the grappling to stop early (instead of directing the player towards the hook, which would cause them to lose e.g. the upward momentum when grappling up a wall). Sliding is now explicit, by keeping the same velocity as when the hook was first reached - before the direction is changed from e.g. hitting a wall or sliding past the hook on the floor.

Also made HookHead and PlayerController/-State not change the state to grappling if already grappling.

Also, removed setting `stopped = true` in HookHead unnecessarily.